requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
bcrypt==4.3.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.104.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2==2.9.10
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.24.0.post1
win32_setctime==1.2.0
